requirements.txt
accelerate==1.10.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
annotated-types==0.7.0
asttokens==3.0.0
async-timeout==5.0.1
attrs==25.3.0
autoawq==0.2.9
autoawq_kernels==0.0.9
boto3==1.40.16
botocore==1.40.16
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.2.1
comm==0.2.3
datasets==4.0.0
debugpy==1.8.16
decorator==5.2.1
dill==0.3.8
exceptiongroup==1.3.0
executing==2.2.0
filelock==3.13.1
frozenlist==1.7.0
fsspec==2024.6.1
hf-xet==1.1.8
huggingface-hub==0.34.4
idna==3.10
importlib_metadata==8.7.0
ipykernel==6.30.1
ipython==8.37.0
jedi==0.19.2
Jinja2==3.1.4
jmespath==1.0.1
jupyter_client==8.6.3
jupyter_core==5.8.1
markdown-it-py==4.0.0
MarkupSafe==2.1.5
matplotlib-inline==0.1.7
mdurl==0.1.2
modelscope==1.29.1
mpmath==1.3.0
multidict==6.6.4
multiprocess==0.70.16
nest_asyncio==1.6.0
networkx==3.3
numpy==2.1.2
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-ml-py==12.575.51
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.9.86
nvidia-nvtx-cu12==12.1.105
packaging==25.0
pandas==2.3.2
parso==0.8.4
peft==0.17.1
pexpect==4.9.0
pickleshare==0.7.5
pillow==11.0.0
pip==25.1
platformdirs==4.3.8
prettytable==3.16.0
prompt_toolkit==3.0.51
propcache==0.3.2
protobuf==6.32.0
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==21.0.0
pydantic==2.11.7
pydantic_core==2.33.2
pyecharts==2.0.8
Pygments==2.19.2
pynvml==12.0.0
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
pyzmq==27.0.2
regex==2025.7.34
requests==2.32.5
rich==13.9.4
s3transfer==0.13.1
safetensors==0.6.2
setuptools==78.1.1
simplejson==3.20.1
six==1.17.0
stack_data==0.6.3
swankit==0.2.4
swanlab==0.6.8
sympy==1.13.1
tokenizers==0.21.4
torch==2.5.1+cu121
torchaudio==2.5.1+cu121
torchvision==0.20.1+cu121
tornado==6.5.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.51.3
triton==3.1.0
typing_extensions==4.12.2
typing-inspection==0.4.1
tzdata==2025.2
urllib3==2.5.0
wcwidth==0.2.13
wheel==0.45.1
wrapt==1.17.3
xxhash==3.5.0
yarl==1.20.1
zipp==3.23.0
zstandard==0.24.0
